Starts the win conditions. Still needs to do the tie case, however need to first to the recieve blank card for the tie case
