+++ main/model/Card.java
@@ -1,6 +1,6 @@
 package main.model;
 
-class Card {
+public class Card {
     private final String suit;
     private final String rank;
