+++ main/model/Player.java
@@ -35,7 +35,6 @@ public class Player {
             }
         }
 
-        // Adjust for Aces if score is over 21
         while (score > 21 && aceCount > 0) {
             score -= 10;
             aceCount--;
